Touch up integration test structure
