+++ tests/integration_tests/helpers/README.md
@@ -1,0 +1,3 @@
+This directory contains helpers for the integration tests. Note that it is
+added to the sys path in `../conftest.py`, such that these modules are
+available as top level imports in the tests.
